requirements.txt
fastapi
uvicorn
starlette
itsdangerous
RPi.GPIO